API classes moved to new bundle.
