+++ extensions/bundles/router.capabilities.api/src/test/resources/ospfService.xml
