+++ extensions/bundles/router.capability.ospf/pom.xml
@@ -30,6 +30,10 @@
 			<groupId>org.apache.karaf.shell</groupId>
 			<artifactId>org.apache.karaf.shell.console</artifactId>
 		</dependency>
+		<dependency>
+		    <groupId>org.opennaas</groupId>
+		  	<artifactId>org.opennaas.extensions.router.capabilities.api</artifactId>  
+		</dependency>
 	</dependencies>
 	<build>
 		<plugins>
